fix: standard image recognition works again

+++ lib/helper/isolate_inference.dart
@@ -61,14 +61,14 @@ class IsolateInference {
       // Set tensor input [1, 224, 224, 3]
       final input = [imageMatrix];
       // Set tensor output [1, 1001]
-      final output = [List<double>.filled(isolateModel.outputShape[1], 0.0)];
+      final output = [List<int>.filled(isolateModel.outputShape[1], 0)];
       // // Run inference
       Interpreter interpreter =
           Interpreter.fromAddress(isolateModel.interpreterAddress);
       interpreter.run(input, output);
       // Get first output tensor
       final result = output.first;
-      double maxScore = result.reduce((a, b) => a + b);
+      int maxScore = result.reduce((a, b) => a + b);
       // Set classification map {label: points}
       var classification = <String, double>{};
       for (var i = 0; i < result.length; i++) {
